Rename ShapeGallery names to Shapes
