+++ Elmanager/LevelEditor/Shapes/CustomShapeControl.resx
