+++ Elmanager/LevelEditor/Shapes/ShapeSelectionForm.resx
